pyproject.toml
[tool.poetry]
name = "renku-data-services"
version = "0.0.1"
description = "Collection of services that deal with data in Postgres."
authors = ["Swiss Data Science Center <contact@datascience.ch>"]
readme = "README.md"
license = "Apache License 2.0"
packages = [
  { include = "renku_data_services/data_api", from = "bases" },
  { include = "renku_data_services/keycloak_sync", from = "bases" },
  { include = "renku_data_services/authn", from = "components" },
  { include = "renku_data_services/db_config", from = "components" },
  { include = "renku_data_services/app_config", from = "components" },
  { include = "renku_data_services/authz", from = "components" },
  { include = "renku_data_services/crc", from = "components" },
  { include = "renku_data_services/crc_migrations", from = "components" },
  { include = "renku_data_services/k8s", from = "components" },
  { include = "renku_data_services/migrations", from = "components" },
  { include = "renku_data_services/errors", from = "components" },
  { include = "renku_data_services/base_models", from = "components" },
  { include = "renku_data_services/base_api", from = "components" },
  { include = "renku_data_services/storage", from = "components" },
  { include = "renku_data_services/storage_migrations", from = "components" },
  { include = "renku_data_services/user_preferences", from = "components" },
  { include = "renku_data_services/user_preferences_migrations", from = "components" },
  { include = "renku_data_services/utils", from = "components" },
  { include = "renku_data_services/git", from = "components" },
  { include = "renku_data_services/users", from = "components" },
  { include = "renku_data_services/users_migrations", from = "components" },
  { include = "renku_data_services/project", from = "components" },
  { include = "renku_data_services/project_migrations", from = "components" },
  { include = "renku_data_services/events_migrations", from = "components" },
  { include = "renku_data_services/message_queue", from = "components" },
  { include = "renku_data_services/namespace", from = "components" },
  { include = "renku_data_services/session", from = "components" },
  { include = "renku_data_services/session_migrations", from = "components" },
  { include = "renku_data_services/events_migrations", from = "components" },
  { include = "renku_data_services/message_queue", from = "components" },
]

[tool.poetry.dependencies]
python = "^3.12"
sanic = { extras = ["ext"], version = "^23.3.0" }
pydantic = {extras = ["email"], version = "^2.6.4"}
datamodel-code-generator = "^0.24.2"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.9" }
alembic = "^1.10.3"
asyncpg = "^0.29.0"
pyjwt = { extras = ["crypto"], version = "^2.6.0" }
tenacity = "^8.2.2"
httpx = "<0.24"
kubernetes = "^26.1.0"
python-ulid = "^1.1.0"
python-gitlab = "^3.15.0"
psycopg = { version = "^3.1.10", extras = ["binary"] }
urllib3 = "^2.2.1"
deepmerge = "^1.1.0"
authlib = "^1.2.1"
redis = "^5.0.1"
dataclasses-avroschema = "^0.57.4"
undictify = "^0.11.3"
types-redis = "^4.6.0.20240218"
prometheus-sanic = "^3.0.0"
sentry-sdk = { version = "^1.43.0", extras = ["sanic"] }
authzed = "^0.14.0"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.5"
mypy = "^1.1.1"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
pre-commit = "^3.2.2"
sanic-testing = "^23.3.0"
aiosqlite = "^0.18.0"
types-pyyaml = "^6.0.12.9"
schemathesis = "^3.19.0"
pytest-asyncio = "^0.21.1"
pytest-postgresql = "^5.0.0"
types-urllib3 = "^1.26.25.14"
pyavro-gen = "^0.3.3"
avro-preprocessor = "^0.3.0"
fakeredis = "^2.21.1"
ruff = "^0.3.4"
debugpy = "^1.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py311"
output-format = "full"
include = ["*.py", "*.pyi"]
exclude = ["*/avro_models/*"]

[tool.ruff.lint]
select = [
  # isort
  "I",
  # pydocstyle
  "D",
  # pyflakes
  "F",
  # pycodestyle
  "E",
  "W",
  # pyupgrade
  "UP",
  # flake8-simplify
  "SIM",
]
ignore = [
  "D105",
  "D107",
  "D202",
  "D211",
  "D213",
  "D401",
  "E202",
  "E226",
  "E231",
  "UP007",
  "UP038"
]

[tool.ruff.lint.per-file-ignores]
"test/*" = ["D"]
"*/crc_migrations/*" = ["D", "I"]
"*/events_migrations/*" = ["D", "I"]
"*/storage_migrations/*" = ["D", "I"]
"*/user_preferences_migrations/*" = ["D", "I"]
"*/project_migrations/*" = ["D", "I"]
"*/session_migrations/*" = ["D", "I"]
"*/users_migrations/*" = ["D", "I"]
"*/versions/*" = ["D", "E", "W"]
"apispec.py" = ["D", "E", "W", "I", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["renku_data_services"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.bandit]
skips = ["B101", "B603", "B607", "B404"]

[tool.pytest.ini_options]
addopts = "--cov components/ --cov bases/ --cov-report=term-missing -v"
doctest_optionflags = "ALLOW_UNICODE"
testpaths = ["bases", "components", "test"]
markers = ["integration: mark a test as a integration."]
filterwarnings = [
  "ignore:<class 'pytest_black.BlackItem'> is not using a cooperative constructor:pytest.PytestDeprecationWarning",
  "ignore:distutils Version classes are deprecated. Use packaging.version instead:DeprecationWarning",
]
postgresql_host = "localhost"
postgresql_user = "renku"
postgresql_password = "renku"

[tool.mypy]
python_version = "3.12"
mypy_path = ["components", "bases"]
files = ["bases/**/*.py", "components/**/*.py"]
namespace_packages = true
explicit_package_bases = true
exclude = ["test"]

pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
warn_unreachable = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
  "renku_data_services.crc.apispec",
  "renku_data_services.storage.apispec",
  "renku_data_services.project.apispec",
  "renku_data_services.session.apispec",
  "renku_data_services.user_preferences.apispec",
  "renku_data_services.users.apispec",
  "renku_data_services.data_api.error_handler",
  "renku_data_services.namespace.apispec",
]
ignore_errors = true

[[tool.mypy.overrides]]
module = [
  "sanic_testing.*",
  "asyncpg.*",
  "kubernetes.*",
  "deepmerge.*",
  "pyavro_gen.*",
  "avro_preprocessor.*",
  "undictify.*",
  "dataclasses_avroschema",
  "urllib3.*",
  "prometheus_sanic.*",
  "authzed.*",
  "grpcutil.*",
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["bases/", "components/"]
concurrency = ["thread", "greenlet"]
omit = ["components/renku_data_services/migrations/resource_pools/*"]
